requirements.txt
Flask==2.3.3
tensorflow==2.13.0
Pillow==10.0.1
matplotlib==3.7.2
numpy==1.24.3
google-generativeai==0.3.2
